Move converters to fzLocal
